@@ -224,7 +224,9 @@ class _$Preparing extends Preparing {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$PreparingToJson(this);
+    return _$$PreparingToJson(
+      this,
+    );
   }
 }
 
@@ -392,7 +394,9 @@ class _$InProgress extends InProgress {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$InProgressToJson(this);
+    return _$$InProgressToJson(
+      this,
+    );
   }
 }
 
@@ -404,8 +408,8 @@ abstract class InProgress extends UploadState {
   factory InProgress.fromJson(Map<String, dynamic> json) =
       _$InProgress.fromJson;
 
-  int get uploaded => throw _privateConstructorUsedError;
-  int get total => throw _privateConstructorUsedError;
+  int get uploaded;
+  int get total;
   @JsonKey(ignore: true)
   _$$InProgressCopyWith<_$InProgress> get copyWith =>
       throw _privateConstructorUsedError;
@@ -531,7 +535,9 @@ class _$Success extends Success {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$SuccessToJson(this);
+    return _$$SuccessToJson(
+      this,
+    );
   }
 }
 
@@ -686,7 +692,9 @@ class _$Failed extends Failed {
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$FailedToJson(this);
+    return _$$FailedToJson(
+      this,
+    );
   }
 }
 
@@ -696,7 +704,7 @@ abstract class Failed extends UploadState {
 
   factory Failed.fromJson(Map<String, dynamic> json) = _$Failed.fromJson;
 
-  String get error => throw _privateConstructorUsedError;
+  String get error;
   @JsonKey(ignore: true)
   _$$FailedCopyWith<_$Failed> get copyWith =>
       throw _privateConstructorUsedError;
