@@ -1,6 +1,7 @@
 import 'package:stream_chat/src/core/platform_detector/platform_detector.dart';
 
 /// Stub implementation
-PlatformType get currentPlatform {
-  throw UnimplementedError();
-}
+PlatformType get currentPlatform => throw UnimplementedError();
+
+/// Stub implementation
+bool get isFlutterTestEnvironment => throw UnimplementedError();
