@@ -1,6 +1,7 @@
 library stream_chat;
 
 export 'package:async/async.dart';
+export 'package:dio/src/cancel_token.dart';
 export 'package:dio/src/dio_error.dart';
 export 'package:dio/src/multipart_file.dart';
 export 'package:dio/src/options.dart';
@@ -9,8 +10,7 @@ export 'package:logging/logging.dart' show Logger, Level, LogRecord;
 export 'package:rate_limiter/rate_limiter.dart';
 export 'package:uuid/uuid.dart';
 
-export './src/core/api/attachment_file_uploader.dart'
-    show AttachmentFileUploader;
+export './src/core/api/attachment_file_uploader.dart';
 export './src/core/api/requests.dart';
 export './src/core/api/requests.dart';
 export './src/core/api/responses.dart';
@@ -40,11 +40,12 @@ export './src/permission_type.dart';
 export './src/ws/connection_status.dart';
 export 'src/client/channel.dart';
 export 'src/client/client.dart';
+export 'src/client/key_stroke_handler.dart';
 export 'src/core/api/attachment_file_uploader.dart' show AttachmentFileUploader;
 export 'src/core/api/requests.dart';
 export 'src/core/api/requests.dart';
 export 'src/core/api/responses.dart';
-export 'src/core/api/stream_chat_api.dart' show PushProvider;
+export 'src/core/api/stream_chat_api.dart';
 export 'src/core/error/error.dart';
 export 'src/core/models/action.dart';
 export 'src/core/models/attachment.dart';
@@ -63,6 +64,7 @@ export 'src/core/models/own_user.dart';
 export 'src/core/models/reaction.dart';
 export 'src/core/models/read.dart';
 export 'src/core/models/user.dart';
+export 'src/core/platform_detector/platform_detector.dart';
 export 'src/core/util/extension.dart';
 export 'src/db/chat_persistence_client.dart';
 export 'src/event_type.dart';
