@@ -1,5 +1,4 @@
 import 'package:flutter/widgets.dart';
-
 import 'package:stream_chat_flutter/src/localization/translations.dart'
     show Translations;
 
@@ -28,9 +27,10 @@ abstract class StreamChatLocalizations implements Translations {
   /// ```dart
   /// tooltip: StreamChatLocalizations.of(context).streamChatLabel,
   /// ```
-  static StreamChatLocalizations? of(BuildContext context) =>
-      Localizations.of<StreamChatLocalizations>(
-        context,
-        StreamChatLocalizations,
-      );
+  static StreamChatLocalizations? of(BuildContext context) {
+    return Localizations.of<StreamChatLocalizations>(
+      context,
+      StreamChatLocalizations,
+    );
+  }
 }
