@@ -2,8 +2,10 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_portal/flutter_portal.dart';
+import 'package:stream_chat_flutter/src/video/vlc/vlc_manager.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+/// {@template streamChat}
 /// Widget used to provide information about the chat to the widget tree
 ///
 /// class MyApp extends StatelessWidget {
@@ -25,19 +27,21 @@ import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 /// }
 ///
 /// Use [StreamChat.of] to get the current [StreamChatState] instance.
+/// {@endtemplate}
 class StreamChat extends StatefulWidget {
-  /// Constructor for creating a [StreamChat] widget
+  /// {@macro streamChat}
   const StreamChat({
     super.key,
     required this.client,
     required this.child,
     this.streamChatThemeData,
+    this.streamChatConfigData,
     this.onBackgroundEventReceived,
     this.backgroundKeepAlive = const Duration(minutes: 1),
     this.connectivityStream,
   });
 
-  /// Client to do chat ops with
+  /// Client to do chat operations with
   final StreamChatClient client;
 
   /// Child which inherits details
@@ -45,6 +49,9 @@ class StreamChat extends StatefulWidget {
 
   /// Theme to pass on
   final StreamChatThemeData? streamChatThemeData;
+
+  /// Non-theme related UI configuration options.
+  final StreamChatConfigurationData? streamChatConfigData;
 
   /// The amount of time that will pass before disconnecting the client
   /// in the background
@@ -84,40 +91,57 @@ class StreamChatState extends State<StreamChat> {
   /// Gets client from widget
   StreamChatClient get client => widget.client;
 
+  /// Gets configuration options from widget
+  StreamChatConfigurationData get streamChatConfigData =>
+      widget.streamChatConfigData ?? StreamChatConfigurationData();
+
+  @override
+  void initState() {
+    super.initState();
+    // Ensures that VLC only initializes in real desktop environments
+    if (!isTestEnvironment && isDesktopVideoPlayerSupported) {
+      VlcManager.instance.initialize();
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     final theme = _getTheme(context, widget.streamChatThemeData);
     return Portal(
-      child: StreamChatTheme(
-        data: theme,
-        child: Builder(
-          builder: (context) {
-            final materialTheme = Theme.of(context);
-            final streamTheme = StreamChatTheme.of(context);
-            return Theme(
-              data: materialTheme.copyWith(
-                primaryIconTheme: streamTheme.primaryIconTheme,
-                colorScheme: materialTheme.colorScheme.copyWith(
-                  secondary: streamTheme.colorTheme.accentPrimary,
+      child: StreamChatConfiguration(
+        data: streamChatConfigData,
+        child: StreamChatTheme(
+          data: theme,
+          child: Builder(
+            builder: (context) {
+              final materialTheme = Theme.of(context);
+              final streamTheme = StreamChatTheme.of(context);
+              return Theme(
+                data: materialTheme.copyWith(
+                  primaryIconTheme: streamTheme.primaryIconTheme,
+                  colorScheme: materialTheme.colorScheme.copyWith(
+                    secondary: streamTheme.colorTheme.accentPrimary,
+                  ),
                 ),
-              ),
-              child: StreamChatCore(
-                client: client,
-                onBackgroundEventReceived: widget.onBackgroundEventReceived,
-                backgroundKeepAlive: widget.backgroundKeepAlive,
-                connectivityStream: widget.connectivityStream,
-                child: Builder(
-                  builder: (context) {
-                    StreamChatClient.additionalHeaders = {
-                      'X-Stream-Client': '${StreamChatClient.defaultUserAgent}-'
-                          'ui-${StreamChatClient.packageVersion}',
-                    };
-                    return widget.child ?? const Offstage();
-                  },
+                child: StreamChatCore(
+                  client: client,
+                  onBackgroundEventReceived: widget.onBackgroundEventReceived,
+                  backgroundKeepAlive: widget.backgroundKeepAlive,
+                  connectivityStream: widget.connectivityStream,
+                  child: Builder(
+                    builder: (context) {
+                      StreamChatClient.additionalHeaders = {
+                        'X-Stream-Client':
+                            '${StreamChatClient.defaultUserAgent}-'
+                                'ui-${StreamChatClient.packageVersion}',
+                      };
+                      return widget.child ?? const Offstage();
+                    },
+                  ),
                 ),
-              ),
-            );
-          },
+              );
+            },
+          ),
         ),
       ),
     );
@@ -140,11 +164,10 @@ class StreamChatState extends State<StreamChat> {
 
   @override
   void didChangeDependencies() {
-    final currentLocale = Localizations.localeOf(context);
-    final languageCode = currentLocale.languageCode;
+    final currentLocale = Localizations.localeOf(context).toString();
     final availableLocales = Jiffy.getAllAvailableLocales();
-    if (availableLocales.contains(languageCode)) {
-      Jiffy.locale(languageCode);
+    if (availableLocales.contains(currentLocale)) {
+      Jiffy.locale(currentLocale);
     }
     super.didChangeDependencies();
   }
