@@ -2,10 +2,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/src/theme/avatar_theme.dart';
 
-/// {@macro message_theme_data}
-@Deprecated("Use 'StreamMessageThemeData' instead")
-typedef MessageThemeData = StreamMessageThemeData;
-
 /// {@template message_theme_data}
 /// Class for getting message theme
 /// {@endtemplate}
@@ -77,57 +73,59 @@ class StreamMessageThemeData with Diagnosticable {
     Color? reactionsBorderColor,
     Color? reactionsMaskColor,
     Color? linkBackgroundColor,
-  }) =>
-      StreamMessageThemeData(
-        messageTextStyle: messageTextStyle ?? this.messageTextStyle,
-        messageAuthorStyle: messageAuthorStyle ?? this.messageAuthorStyle,
-        messageLinksStyle: messageLinksStyle ?? this.messageLinksStyle,
-        createdAtStyle: createdAtStyle ?? this.createdAtStyle,
-        messageBackgroundColor:
-            messageBackgroundColor ?? this.messageBackgroundColor,
-        messageBorderColor: messageBorderColor ?? this.messageBorderColor,
-        avatarTheme: avatarTheme ?? this.avatarTheme,
-        repliesStyle: repliesStyle ?? this.repliesStyle,
-        reactionsBackgroundColor:
-            reactionsBackgroundColor ?? this.reactionsBackgroundColor,
-        reactionsBorderColor: reactionsBorderColor ?? this.reactionsBorderColor,
-        reactionsMaskColor: reactionsMaskColor ?? this.reactionsMaskColor,
-        linkBackgroundColor: linkBackgroundColor ?? this.linkBackgroundColor,
-      );
+  }) {
+    return StreamMessageThemeData(
+      messageTextStyle: messageTextStyle ?? this.messageTextStyle,
+      messageAuthorStyle: messageAuthorStyle ?? this.messageAuthorStyle,
+      messageLinksStyle: messageLinksStyle ?? this.messageLinksStyle,
+      createdAtStyle: createdAtStyle ?? this.createdAtStyle,
+      messageBackgroundColor:
+          messageBackgroundColor ?? this.messageBackgroundColor,
+      messageBorderColor: messageBorderColor ?? this.messageBorderColor,
+      avatarTheme: avatarTheme ?? this.avatarTheme,
+      repliesStyle: repliesStyle ?? this.repliesStyle,
+      reactionsBackgroundColor:
+          reactionsBackgroundColor ?? this.reactionsBackgroundColor,
+      reactionsBorderColor: reactionsBorderColor ?? this.reactionsBorderColor,
+      reactionsMaskColor: reactionsMaskColor ?? this.reactionsMaskColor,
+      linkBackgroundColor: linkBackgroundColor ?? this.linkBackgroundColor,
+    );
+  }
 
   /// Linearly interpolate from one [StreamMessageThemeData] to another.
   StreamMessageThemeData lerp(
     StreamMessageThemeData a,
     StreamMessageThemeData b,
     double t,
-  ) =>
-      StreamMessageThemeData(
-        avatarTheme: const StreamAvatarThemeData()
-            .lerp(a.avatarTheme!, b.avatarTheme!, t),
-        createdAtStyle: TextStyle.lerp(a.createdAtStyle, b.createdAtStyle, t),
-        messageAuthorStyle:
-            TextStyle.lerp(a.messageAuthorStyle, b.messageAuthorStyle, t),
-        messageBackgroundColor:
-            Color.lerp(a.messageBackgroundColor, b.messageBackgroundColor, t),
-        messageBorderColor:
-            Color.lerp(a.messageBorderColor, b.messageBorderColor, t),
-        messageLinksStyle:
-            TextStyle.lerp(a.messageLinksStyle, b.messageLinksStyle, t),
-        messageTextStyle:
-            TextStyle.lerp(a.messageTextStyle, b.messageTextStyle, t),
-        reactionsBackgroundColor: Color.lerp(
-          a.reactionsBackgroundColor,
-          b.reactionsBackgroundColor,
-          t,
-        ),
-        reactionsBorderColor:
-            Color.lerp(a.messageBorderColor, b.reactionsBorderColor, t),
-        reactionsMaskColor:
-            Color.lerp(a.reactionsMaskColor, b.reactionsMaskColor, t),
-        repliesStyle: TextStyle.lerp(a.repliesStyle, b.repliesStyle, t),
-        linkBackgroundColor:
-            Color.lerp(a.linkBackgroundColor, b.linkBackgroundColor, t),
-      );
+  ) {
+    return StreamMessageThemeData(
+      avatarTheme:
+          const StreamAvatarThemeData().lerp(a.avatarTheme!, b.avatarTheme!, t),
+      createdAtStyle: TextStyle.lerp(a.createdAtStyle, b.createdAtStyle, t),
+      messageAuthorStyle:
+          TextStyle.lerp(a.messageAuthorStyle, b.messageAuthorStyle, t),
+      messageBackgroundColor:
+          Color.lerp(a.messageBackgroundColor, b.messageBackgroundColor, t),
+      messageBorderColor:
+          Color.lerp(a.messageBorderColor, b.messageBorderColor, t),
+      messageLinksStyle:
+          TextStyle.lerp(a.messageLinksStyle, b.messageLinksStyle, t),
+      messageTextStyle:
+          TextStyle.lerp(a.messageTextStyle, b.messageTextStyle, t),
+      reactionsBackgroundColor: Color.lerp(
+        a.reactionsBackgroundColor,
+        b.reactionsBackgroundColor,
+        t,
+      ),
+      reactionsBorderColor:
+          Color.lerp(a.messageBorderColor, b.reactionsBorderColor, t),
+      reactionsMaskColor:
+          Color.lerp(a.reactionsMaskColor, b.reactionsMaskColor, t),
+      repliesStyle: TextStyle.lerp(a.repliesStyle, b.repliesStyle, t),
+      linkBackgroundColor:
+          Color.lerp(a.linkBackgroundColor, b.linkBackgroundColor, t),
+    );
+  }
 
   /// Merge with a theme
   StreamMessageThemeData merge(StreamMessageThemeData? other) {
