@@ -1,77 +1,100 @@
 export 'package:jiffy/jiffy.dart';
+export 'package:photo_manager/photo_manager.dart'
+    show ThumbnailSize, ThumbnailFormat;
+export 'package:stream_chat_flutter/src/message_widget/parse_attachments.dart';
+export 'package:stream_chat_flutter/src/message_widget/quoted_message.dart';
 export 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 export 'src/attachment/attachment.dart';
-export 'src/attachment_actions_modal.dart';
-export 'src/back_button.dart';
-export 'src/channel_avatar.dart';
-export 'src/channel_header.dart';
-export 'src/channel_list_header.dart';
-export 'src/channel_list_view.dart';
-export 'src/channel_name.dart';
-export 'src/channel_preview.dart';
-export 'src/connection_status_builder.dart';
-export 'src/date_divider.dart';
-export 'src/deleted_message.dart';
-export 'src/extension.dart' show IconButtonX;
-export 'src/full_screen_media.dart';
-export 'src/gallery_footer.dart';
-export 'src/gallery_header.dart';
-export 'src/gradient_avatar.dart';
-export 'src/info_tile.dart';
+export 'src/attachment/attachment_title.dart';
+export 'src/attachment/handler/stream_attachment_handler.dart';
+export 'src/attachment/image_attachment.dart';
+export 'src/attachment/image_group.dart';
+export 'src/attachment/stream_attachment_package.dart';
+export 'src/attachment/url_attachment.dart';
+export 'src/attachment/video_attachment.dart';
+export 'src/attachment_actions_modal/attachment_actions_modal.dart';
+export 'src/autocomplete/stream_autocomplete.dart';
+export 'src/avatars/gradient_avatar.dart';
+export 'src/avatars/group_avatar.dart';
+export 'src/avatars/user_avatar.dart';
+export 'src/bottom_sheets/attachment_modal_sheet.dart';
+export 'src/bottom_sheets/edit_message_sheet.dart';
+export 'src/bottom_sheets/error_alert_sheet.dart';
+export 'src/bottom_sheets/stream_channel_info_bottom_sheet.dart';
+export 'src/channel/channel_header.dart';
+export 'src/channel/channel_info.dart';
+export 'src/channel/channel_list_header.dart';
+export 'src/channel/channel_name.dart';
+export 'src/channel/channel_preview.dart';
+export 'src/channel/stream_channel_avatar.dart';
+export 'src/channel/stream_channel_name.dart';
+export 'src/channel/stream_message_preview_text.dart';
+export 'src/fullscreen_media/fsm_enums.dart';
+export 'src/fullscreen_media/full_screen_media.dart';
+export 'src/fullscreen_media/full_screen_media_builder.dart';
+export 'src/gallery/gallery_footer.dart';
+export 'src/gallery/gallery_header.dart';
+export 'src/indicators/sending_indicator.dart';
+export 'src/indicators/typing_indicator.dart';
+export 'src/indicators/unread_indicator.dart';
+export 'src/indicators/upload_progress_indicator.dart';
+export 'src/keyboard_shortcuts/keyboard_shortcut_runner.dart';
 export 'src/localization/stream_chat_localizations.dart';
 export 'src/localization/translations.dart' show DefaultTranslations;
-export 'src/message_action.dart';
-// ignore: deprecated_member_use_from_same_package
-export 'src/message_input.dart' show MessageInput, MessageInputState;
-export 'src/message_list_view.dart';
-export 'src/message_search_item.dart';
-export 'src/message_search_list_view.dart';
-export 'src/message_text.dart';
-export 'src/message_widget.dart';
-export 'src/multi_overlay.dart';
-export 'src/option_list_tile.dart';
-export 'src/reaction_icon.dart';
-export 'src/reaction_picker.dart';
-export 'src/sending_indicator.dart';
-export 'src/stream_attachment_package.dart';
+export 'src/message_actions_modal/message_action.dart';
+export 'src/message_input/attachment_picker/stream_attachment_picker.dart';
+export 'src/message_input/attachment_picker/stream_attachment_picker_bottom_sheet.dart';
+export 'src/message_input/countdown_button.dart';
+export 'src/message_input/enums.dart';
+export 'src/message_input/stream_message_input.dart';
+export 'src/message_input/stream_message_send_button.dart';
+export 'src/message_input/stream_message_text_field.dart';
+export 'src/message_list_view/message_details.dart';
+export 'src/message_list_view/message_list_view.dart';
+export 'src/message_widget/deleted_message.dart';
+export 'src/message_widget/message_text.dart';
+export 'src/message_widget/message_widget.dart';
+export 'src/message_widget/reactions/reaction_picker.dart';
+export 'src/message_widget/text_bubble.dart';
+export 'src/misc/back_button.dart';
+export 'src/misc/connection_status_builder.dart';
+export 'src/misc/date_divider.dart';
+export 'src/misc/info_tile.dart';
+export 'src/misc/option_list_tile.dart';
+export 'src/misc/reaction_icon.dart';
+export 'src/misc/stream_neumorphic_button.dart';
+export 'src/misc/stream_svg_icon.dart';
+export 'src/misc/system_message.dart';
+export 'src/misc/thread_header.dart';
+export 'src/misc/visible_footnote.dart';
+export 'src/scroll_view/channel_scroll_view/stream_channel_grid_tile.dart';
+export 'src/scroll_view/channel_scroll_view/stream_channel_grid_view.dart';
+export 'src/scroll_view/channel_scroll_view/stream_channel_list_tile.dart';
+export 'src/scroll_view/channel_scroll_view/stream_channel_list_view.dart';
+export 'src/scroll_view/member_scroll_view/stream_member_grid_view.dart';
+export 'src/scroll_view/member_scroll_view/stream_member_list_view.dart';
+export 'src/scroll_view/message_search_scroll_view/stream_message_search_grid_view.dart';
+export 'src/scroll_view/message_search_scroll_view/stream_message_search_list_tile.dart';
+export 'src/scroll_view/message_search_scroll_view/stream_message_search_list_view.dart';
+export 'src/scroll_view/photo_gallery/stream_photo_gallery.dart';
+export 'src/scroll_view/photo_gallery/stream_photo_gallery_controller.dart';
+export 'src/scroll_view/photo_gallery/stream_photo_gallery_tile.dart';
+export 'src/scroll_view/stream_scroll_view_empty_widget.dart';
+export 'src/scroll_view/stream_scroll_view_indexed_widget_builder.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_grid_tile.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_grid_tile.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_grid_view.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_grid_view.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_list_tile.dart';
+export 'src/scroll_view/user_scroll_view/stream_user_list_view.dart';
 export 'src/stream_chat.dart';
-export 'src/stream_chat_theme.dart';
-export 'src/stream_neumorphic_button.dart';
-export 'src/stream_svg_icon.dart';
-export 'src/system_message.dart';
+export 'src/stream_chat_configuration.dart';
+export 'src/theme/stream_chat_theme.dart';
 export 'src/theme/themes.dart';
-export 'src/thread_header.dart';
-export 'src/typing_indicator.dart';
-export 'src/unread_indicator.dart';
-export 'src/user_avatar.dart';
-export 'src/user_item.dart';
-export 'src/user_list_view.dart';
-export 'src/user_mention_tile.dart';
-export 'src/utils.dart';
-// v4
-export 'src/v4/message_input/countdown_button.dart';
-export 'src/v4/message_input/stream_attachment_picker.dart';
-export 'src/v4/message_input/stream_message_input.dart';
-export 'src/v4/message_input/stream_message_send_button.dart';
-export 'src/v4/message_input/stream_message_text_field.dart';
-export 'src/v4/scroll_view/channel_scroll_view/stream_channel_grid_tile.dart';
-export 'src/v4/scroll_view/channel_scroll_view/stream_channel_grid_view.dart';
-export 'src/v4/scroll_view/channel_scroll_view/stream_channel_list_tile.dart';
-export 'src/v4/scroll_view/channel_scroll_view/stream_channel_list_view.dart';
-export 'src/v4/scroll_view/message_search_scroll_view/stream_message_search_grid_view.dart';
-export 'src/v4/scroll_view/message_search_scroll_view/stream_message_search_list_tile.dart';
-export 'src/v4/scroll_view/message_search_scroll_view/stream_message_search_list_view.dart';
-export 'src/v4/scroll_view/stream_scroll_view_empty_widget.dart';
-export 'src/v4/scroll_view/stream_scroll_view_indexed_widget_builder.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_grid_tile.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_grid_tile.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_grid_view.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_grid_view.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_list_tile.dart';
-export 'src/v4/scroll_view/user_scroll_view/stream_user_list_view.dart';
-export 'src/v4/stream_channel_avatar.dart';
-export 'src/v4/stream_channel_info_bottom_sheet.dart';
-export 'src/v4/stream_channel_name.dart';
-export 'src/v4/stream_message_preview_text.dart';
-export 'src/visible_footnote.dart';
+export 'src/user/user_item.dart';
+export 'src/user/user_mention_tile.dart';
+export 'src/utils/device_segmentation.dart';
+export 'src/utils/extensions.dart' show IconButtonX;
+export 'src/utils/helpers.dart';
+export 'src/utils/typedefs.dart';
