@@ -11,7 +11,9 @@ import 'package:stream_chat_flutter_core/src/message_text_field_controller.dart'
 /// Pass in a [StreamMessageInputController] as the `valueListenable`.
 typedef StreamMessageValueListenableBuilder = ValueListenableBuilder<Message>;
 
+/// {@template stream_chat_flutter.StreamMessageInputController}
 /// Controller for storing and mutating a [Message] value.
+/// {@endtemplate}
 class StreamMessageInputController extends ValueNotifier<Message> {
   /// Creates a controller for an editable text field.
   ///
@@ -50,111 +52,121 @@ class StreamMessageInputController extends ValueNotifier<Message> {
   StreamMessageInputController._({
     required Message initialMessage,
     Map<RegExp, TextStyleBuilder>? textPatternStyle,
-  })  : _textEditingController = MessageTextFieldController.fromValue(
-          initialMessage.text == null
-              ? TextEditingValue.empty
-              : TextEditingValue(
-                  text: initialMessage.text!,
-                  composing: TextRange.collapsed(initialMessage.text!.length),
-                ),
+  })  : _initialMessage = initialMessage,
+        _textFieldController = MessageTextFieldController.fromValue(
+          _textEditingValueFromMessage(initialMessage),
           textPatternStyle: textPatternStyle,
         ),
-        _initialMessage = initialMessage,
         super(initialMessage) {
-    addListener(_textEditingSyncer);
+    _textFieldController.addListener(_textFieldListener);
   }
 
-  void _textEditingSyncer() {
-    final cleanText = value.command == null
-        ? value.text
-        : value.text?.replaceFirst('/${value.command} ', '');
+  /// Returns the controller of the text field linked to this controller.
+  MessageTextFieldController get textFieldController => _textFieldController;
+  MessageTextFieldController _textFieldController;
 
-    if (cleanText != _textEditingController.text) {
-      final previousOffset = _textEditingController.value.selection.start;
-      final previousText = _textEditingController.text;
-      final diff = (cleanText?.length ?? 0) - previousText.length;
-      _textEditingController
-        ..text = cleanText ?? ''
-        ..selection = TextSelection.collapsed(
-          offset: previousOffset + diff,
-        );
+  Message _initialMessage;
+
+  static TextEditingValue _textEditingValueFromMessage(Message message) {
+    final messageText = message.text;
+    var textEditingValue = TextEditingValue.empty;
+    if (messageText != null) {
+      textEditingValue = TextEditingValue(
+        text: messageText,
+        selection: TextSelection.collapsed(offset: messageText.length),
+      );
     }
+    return textEditingValue;
+  }
+
+  void _textFieldListener() {
+    final text = _textFieldController.text;
+    message = message.copyWith(text: text);
   }
 
   /// Returns the current message associated with this controller.
   Message get message => value;
 
-  /// Returns the controller of the text field linked to this controller.
-  MessageTextFieldController get textEditingController =>
-      _textEditingController;
-  final MessageTextFieldController _textEditingController;
+  /// Sets the current message associated with this controller.
+  set message(Message message) => value = message;
 
-  /// Returns the text of the message.
-  String get text => _textEditingController.text;
+  @override
+  set value(Message message) {
+    super.value = message;
 
-  Message _initialMessage;
-
-  /// Sets the message.
-  set message(Message message) {
-    value = message;
+    // Update text field controller only if message text has changed.
+    final messageText = message.text;
+    final textFieldText = _textFieldController.text;
+    if (messageText != textFieldText) {
+      textEditingValue = _textEditingValueFromMessage(message);
+    }
   }
 
-  /// Sets the message that's being quoted.
-  set quotedMessage(Message message) {
-    value = value.copyWith(
-      quotedMessage: message,
-      quotedMessageId: message.id,
+  /// Text of the message.
+  String get text => _textFieldController.text;
+
+  /// Sets the text of the message.
+  set text(String text) {
+    _textFieldController.text = text;
+  }
+
+  /// The currently selected [text].
+  ///
+  /// If the selection is collapsed, then this property gives the offset of the
+  /// cursor within the text.
+  TextSelection get selection => _textFieldController.selection;
+
+  set selection(TextSelection newSelection) {
+    _textFieldController.selection = selection;
+  }
+
+  /// Returns the textEditingValue associated with this controller.
+  TextEditingValue get textEditingValue => _textFieldController.value;
+
+  set textEditingValue(TextEditingValue value) {
+    _textFieldController.value = value;
+  }
+
+  set quotedMessage(Message quotedMessage) {
+    message = message.copyWith(
+      quotedMessage: quotedMessage,
+      quotedMessageId: quotedMessage.id,
     );
   }
 
   /// Clears the quoted message.
   void clearQuotedMessage() {
-    value = value.copyWith(
+    message = message.copyWith(
       quotedMessageId: null,
       quotedMessage: null,
     );
   }
 
   /// Sets a command for the message.
-  set command(Command command) {
-    value = value.copyWith(
-      command: command.name,
-      text: '/${command.name} ',
+  set command(String command) {
+    // Setting the command should also clear the text and attachments.
+    message = message.copyWith(
+      text: '',
+      attachments: [],
+      command: command,
     );
   }
 
-  /// Sets the text of the message.
-  set text(String newText) {
-    var newTextWithCommand = newText;
-    if (value.command != null) {
-      if (!newText.startsWith('/${value.command}')) {
-        newTextWithCommand = '/${value.command} $newText';
-      }
-    }
-    value = value.copyWith(text: newTextWithCommand);
-  }
-
-  /// Returns the baseOffset of the text field.
-  int get baseOffset => textEditingController.selection.baseOffset;
-
-  /// Returns the start of the selection of the text field.
-  int get selectionStart => textEditingController.selection.start;
-
   /// Sets the [showInChannel] flag of the message.
   set showInChannel(bool newValue) {
-    value = value.copyWith(showInChannel: newValue);
+    message = message.copyWith(showInChannel: newValue);
   }
 
   /// Returns true if the message is in a thread and
   /// should be shown in the main channel as well.
-  bool get showInChannel => value.showInChannel ?? false;
+  bool get showInChannel => message.showInChannel ?? false;
 
   /// Returns the attachments of the message.
-  List<Attachment> get attachments => value.attachments;
+  List<Attachment> get attachments => message.attachments;
 
   /// Sets the list of [attachments] for the message.
   set attachments(List<Attachment> attachments) {
-    value = value.copyWith(attachments: attachments);
+    message = message.copyWith(attachments: attachments);
   }
 
   /// Adds a new attachment to the message.
@@ -212,11 +224,11 @@ class StreamMessageInputController extends ValueNotifier<Message> {
   }
 
   /// Returns the list of mentioned users in the message.
-  List<User> get mentionedUsers => value.mentionedUsers;
+  List<User> get mentionedUsers => message.mentionedUsers;
 
   /// Sets the mentioned users.
   set mentionedUsers(List<User> users) {
-    value = value.copyWith(mentionedUsers: users);
+    message = message.copyWith(mentionedUsers: users);
   }
 
   /// Adds a user to the list of mentioned users.
@@ -239,7 +251,7 @@ class StreamMessageInputController extends ValueNotifier<Message> {
     mentionedUsers = [];
   }
 
-  /// Sets the [message], or [value], to empty.
+  /// Sets the [message], to empty.
   ///
   /// After calling this function, [text], [attachments] and [mentionedUsers]
   /// will all be empty.
@@ -250,23 +262,24 @@ class StreamMessageInputController extends ValueNotifier<Message> {
   /// this method should only be called between frames, e.g. in response to user
   /// actions, not during the build, layout, or paint phases.
   void clear() {
-    value = Message();
-    _textEditingController.clear();
+    message = Message();
   }
 
-  /// Sets the [value] to the initial [Message] value.
+  /// Sets the [message] to the initial [Message] value.
   void reset({bool resetId = true}) {
     if (resetId) {
       final newId = const Uuid().v4();
       _initialMessage = _initialMessage.copyWith(id: newId);
     }
-    value = _initialMessage;
+    // Reset the message to the initial value.
+    message = _initialMessage;
   }
 
   @override
   void dispose() {
-    removeListener(_textEditingSyncer);
-    _textEditingController.dispose();
+    _textFieldController
+      ..removeListener(_textFieldListener)
+      ..dispose();
     super.dispose();
   }
 }
@@ -276,7 +289,7 @@ class StreamMessageInputController extends ValueNotifier<Message> {
 ///
 /// The [StreamMessageInputController] is accessible via the [value] getter.
 /// During state restoration,
-/// the property will restore [StreamMessageInputController.value]
+/// the property will restore [StreamMessageInputController.message]
 /// to the value it had when the restoration data it is getting restored from
 /// was collected.
 class StreamRestorableMessageInputController
@@ -306,5 +319,5 @@ class StreamRestorableMessageInputController
   }
 
   @override
-  String toPrimitives() => json.encode(value.value);
+  String toPrimitives() => json.encode(value.message);
 }
