@@ -4,12 +4,8 @@ export 'package:connectivity_plus/connectivity_plus.dart';
 export 'package:stream_chat/stream_chat.dart';
 
 export 'src/better_stream_builder.dart';
-export 'src/channel_list_core.dart' hide ChannelListCoreState;
-export 'src/channels_bloc.dart';
 export 'src/lazy_load_scroll_view.dart';
 export 'src/message_list_core.dart' hide MessageListCoreState;
-export 'src/message_search_bloc.dart';
-export 'src/message_search_list_core.dart' hide MessageSearchListCoreState;
 export 'src/message_text_field_controller.dart';
 export 'src/paged_value_notifier.dart'
     show PagedValueListenableBuilder, PagedValue, PagedValueNotifier;
@@ -18,9 +14,8 @@ export 'src/stream_channel.dart';
 export 'src/stream_channel_list_controller.dart';
 export 'src/stream_channel_list_event_handler.dart';
 export 'src/stream_chat_core.dart';
+export 'src/stream_member_list_controller.dart';
 export 'src/stream_message_input_controller.dart';
 export 'src/stream_message_search_list_controller.dart';
 export 'src/stream_user_list_controller.dart';
 export 'src/typedef.dart';
-export 'src/user_list_core.dart' hide UserListCoreState;
-export 'src/users_bloc.dart';
