@@ -3,23 +3,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
-part 'stream_chat_localizations_es.dart';
-
-part 'stream_chat_localizations_en.dart';
-
-part 'stream_chat_localizations_fr.dart';
-
-part 'stream_chat_localizations_it.dart';
-
-part 'stream_chat_localizations_ja.dart';
-
-part 'stream_chat_localizations_ko.dart';
-
-part 'stream_chat_localizations_hi.dart';
-
-part 'stream_chat_localizations_pt.dart';
-
 part 'stream_chat_localizations_de.dart';
+part 'stream_chat_localizations_en.dart';
+part 'stream_chat_localizations_es.dart';
+part 'stream_chat_localizations_fr.dart';
+part 'stream_chat_localizations_hi.dart';
+part 'stream_chat_localizations_it.dart';
+part 'stream_chat_localizations_ja.dart';
+part 'stream_chat_localizations_ko.dart';
+part 'stream_chat_localizations_pt.dart';
+part 'stream_chat_localizations_no.dart';
 
 /// The set of supported languages, as language code strings.
 ///
@@ -39,6 +32,7 @@ const kStreamChatSupportedLanguages = {
   'ko',
   'pt',
   'de',
+  'no',
 };
 
 /// Creates a [GlobalStreamChatLocalizations] instance for the given `locale`.
@@ -79,6 +73,8 @@ GlobalStreamChatLocalizations? getStreamChatTranslation(Locale locale) {
       return const StreamChatLocalizationsPt();
     case 'de':
       return const StreamChatLocalizationsDe();
+    case 'no':
+      return const StreamChatLocalizationsNo();
     default:
       return null;
   }
