@@ -124,7 +124,7 @@ class StreamChatLocalizationsJa extends GlobalStreamChatLocalizations {
       'ファイルが大きすぎてアップロードできません。ファイルサイズの制限は${limitInMB}MBです。';
 
   @override
-  String emojiMatchingQueryText(String query) => '「"$query"」とお揃いの絵文字';
+  String get couldNotReadBytesFromFileError => 'ファイルからバイトを読み取れませんでした';
 
   @override
   String get addAFileLabel => 'ファイルの追加';
@@ -361,6 +361,54 @@ class StreamChatLocalizationsJa extends GlobalStreamChatLocalizations {
   ''';
 
   @override
+  String get downloadLabel => 'ダウンロード';
+
+  @override
+  String toggleMuteUnmuteUserText({required bool isMuted}) {
+    if (isMuted) {
+      return 'ユーザーのミュートを解除する';
+    } else {
+      return 'ユーザーをミュート';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteGroupQuestion({required bool isMuted}) {
+    if (isMuted) {
+      return 'このグループのミュートを解除してもよろしいですか？';
+    } else {
+      return 'このグループをミュートしてもよろしいですか？';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteUserQuestion({required bool isMuted}) {
+    if (isMuted) {
+      return 'このユーザーのミュートを解除してもよろしいですか？';
+    } else {
+      return 'このユーザーをミュートしてもよろしいですか？';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteAction({required bool isMuted}) {
+    if (isMuted) {
+      return 'ミュートを解除する';
+    } else {
+      return 'ミュート';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteGroupText({required bool isMuted}) {
+    if (isMuted) {
+      return 'グループのミュートを解除';
+    } else {
+      return 'ミュートグループ';
+    }
+  }
+
+  @override
   String get linkDisabledDetails => 'この会話では、リンクの送信は許可されていません。';
 
   @override
@@ -373,4 +421,11 @@ class StreamChatLocalizationsJa extends GlobalStreamChatLocalizations {
     }
     return '$unreadCountつの未読メッセージ';
   }
+
+  @override
+  String get enableFileAccessMessage =>
+      '友達と共有できるように、' '\nファイルへのアクセスを有効にしてください。';
+
+  @override
+  String get allowFileAccessMessage => 'ファイルへのアクセスを許可する';
 }
