@@ -126,8 +126,8 @@ class StreamChatLocalizationsPt extends GlobalStreamChatLocalizations {
       'O tamanho máximo dos arquivos é de $limitInMB MB.';
 
   @override
-  String emojiMatchingQueryText(String query) =>
-      'Emoji correspondente a "$query"';
+  String get couldNotReadBytesFromFileError =>
+      'Não foi possível ler os bytes do arquivo.';
 
   @override
   String get addAFileLabel => 'Adicionar um arquivo';
@@ -372,6 +372,54 @@ Não é possível adicionar mais de $limit arquivos de uma vez
   String get slowModeOnLabel => 'Modo lento ativado';
 
   @override
+  String get downloadLabel => 'Download';
+
+  @override
+  String toggleMuteUnmuteUserText({required bool isMuted}) {
+    if (isMuted) {
+      return 'Ativar o som do usuário';
+    } else {
+      return 'Silenciar usuário';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteGroupQuestion({required bool isMuted}) {
+    if (isMuted) {
+      return 'Tem certeza de que deseja ativar o som deste grupo?';
+    } else {
+      return 'Tem certeza de que deseja silenciar este grupo?';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteUserQuestion({required bool isMuted}) {
+    if (isMuted) {
+      return 'Tem certeza de que deseja ativar o som deste usuário?';
+    } else {
+      return 'Tem certeza de que deseja silenciar este usuário?';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteAction({required bool isMuted}) {
+    if (isMuted) {
+      return 'ATIVAR MUDO';
+    } else {
+      return 'MUDO';
+    }
+  }
+
+  @override
+  String toggleMuteUnmuteGroupText({required bool isMuted}) {
+    if (isMuted) {
+      return 'Reativar o som do grupo';
+    } else {
+      return 'Silenciar Grupo';
+    }
+  }
+
+  @override
   String get linkDisabledDetails =>
       'O envio de links não é permitido nesta conversa.';
 
@@ -392,4 +440,11 @@ Não é possível adicionar mais de $limit arquivos de uma vez
     }
     return '$unreadCount mensagens não lidas';
   }
+
+  @override
+  String get enableFileAccessMessage =>
+      'Ative o acesso aos arquivos' '\npara poder compartilhá-los com amigos.';
+
+  @override
+  String get allowFileAccessMessage => 'Permitir acesso aos arquivos';
 }
