@@ -63,7 +63,6 @@ void main() {
       expect(localizations.instantCommandsLabel, isNotNull);
       expect(localizations.fileTooLargeAfterCompressionError(33), isNotNull);
       expect(localizations.fileTooLargeError(33), isNotNull);
-      expect(localizations.emojiMatchingQueryText('sahil'), isNotNull);
       expect(localizations.addAFileLabel, isNotNull);
       expect(localizations.photoFromCameraLabel, isNotNull);
       expect(localizations.uploadAFileLabel, isNotNull);
@@ -178,6 +177,27 @@ void main() {
       expect(localizations.fileText, isNotNull);
       expect(localizations.replyToMessageLabel, isNotNull);
       expect(localizations.attachmentLimitExceedError(3), isNotNull);
+      expect(
+        localizations.galleryPaginationText(currentPage: 1, totalPages: 2),
+        isNotNull,
+      );
+      expect(localizations.slowModeOnLabel, isNotNull);
+      expect(localizations.linkDisabledDetails, isNotNull);
+      expect(localizations.linkDisabledError, isNotNull);
+      expect(localizations.sendMessagePermissionError, isNotNull);
+      expect(localizations.couldNotReadBytesFromFileError, isNotNull);
+      expect(localizations.toggleMuteUnmuteAction(isMuted: false), isNotNull);
+      expect(localizations.downloadLabel, isNotNull);
+      expect(localizations.toggleMuteUnmuteGroupQuestion(isMuted: true),
+          isNotNull);
+      expect(localizations.toggleMuteUnmuteGroupText(isMuted: true), isNotNull);
+      expect(
+          localizations.toggleMuteUnmuteUserQuestion(isMuted: true), isNotNull);
+      expect(localizations.toggleMuteUnmuteUserText(isMuted: true), isNotNull);
+      expect(localizations.viewLibrary, isNotNull);
+      expect(localizations.unreadMessagesSeparatorText(2), isNotNull);
+      expect(localizations.enableFileAccessMessage, isNotNull);
+      expect(localizations.allowFileAccessMessage, isNotNull);
     });
   }
 
